M rtmp-opt.sh

+++ rtmp-opt.sh
@@ -110,11 +110,12 @@ done
 
 fd ()
 {
+  [ -a kk ] || exit
   printf '\n\n'
   fold -w69 kk
   rm a.flv kk
 }
 
-trap fd 0 INT
+trap fd 0 2
 echo ${bb[@]} -o a.flv > kk
 eval ${bb[@]} -o a.flv |& tee -a kk
